Allow viewport stretch without shrink
Shrink can now go down to zero, and gets special treatment.
Code that forces the viewport resolution checks that shrink is greater than zero.
Transforms are done by dividing the sizes of the viewport and the container (not necessarily in that order), instead of using the shrink value directly.
To accomplish this, _send_event_to_viewports does the scaling now, instead of _gui_input.
I always cast viewport size to floats, because otherwise it casts it implicitly, but only when it's on the right of the operator, and I don't like playing dice.